backend/requirements.txt
supabase>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
playwright>=1.40.0
fake-useragent>=1.4.0
